+++ packages/amplify-schema-validator/src/__tests__/schemas/invalid-index-is-defined-once-in-model.graphql
