+++ packages/amplify-schema-validator/src/__tests__/schemas/valid-index-is-defined-once-in-model.graphql
