+++ packages/amplify-schema-validator/src/index.ts
@@ -9,7 +9,7 @@ import { validateFieldsMatchInParentModel } from './validators/fields-match-in-p
 import { ValidationError } from './exceptions/validation-error';
 import { validateFieldIsDefinedOnce } from './validators/field-is-defined-once';
 import { validatetypeIsDefinedOnce } from './validators/type-is-defined-once';
-import { validateIndexIsDefinedOnce } from './validators/one-index-with-same-name';
+import { validateIndexIsDefinedOnce } from './validators/index-is-defined-once-in-model';
 import { validateIndexExistsInRelatedModel } from './validators/index-exists-in-related-model';
 import { validateEnumIsDefinedOnce } from './validators/enum-is-defined-once';
 import { validateKeyExistsInRelatedModel } from './validators/key-exists-in-related-model';
